refactor: 音楽再生の仕組みのJSを別ファイルに切り出した(#5)
